Make the stand-alone test script mimic tox
The tox setup does not work on Ubuntu Precise for some reason,
but running nose directly does. This update tweaks the stand-alone
script run_tests.sh to mimic the behavior of tox. It also
adds the "cover" directory created by the coverage plugin to nose
to the list of files being ignored by git.

Change-Id: I65224cf4bf75e6b7fc0207b3545c466bdac5a35c