refactor meter event publishing code
There were two places that had all of the logic for constructing
a metering message and sending it to the appropriate exchanges.
This changeset combines those into one function and invokes
the function from the old locations.

Update tox.ini to work around an issue with providing access to
the global site-packages (https://bitbucket.org/hpk42/tox/issue/32).

Change-Id: If2b01edbc0c372907a04baf69f1137575b6921bd