Add indexes to MongoDB driver
Add some compound key indexes for the typical queries
run by the engine. Separate variations are used for
queries with user_id and project_id because they are
usually mutually exclusive and because of the way the
indexes are computed and stored the database won't take
advantage of an index using both.

bug 1021324

Change-Id: I9d6bbd1e7ac5a633fa3ead131a4ebc0a83226633
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>