Change compute pollster API to work on one instance at a time
We will eventually need the compute agent to accept instructions to
poll the status of an instance before nova deletes it (for details
see bug 1005944). This change prepares for that fix by making
the compute agent pollsters take an instance as an argument
provided by the caller, instead of accessing the database directly
to look up the instances known to be running on the current
host.

Change-Id: I6a16405fd65bfb3c190d02f3d70b2bde7fc0fc83
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>