Add API endpoint for listing raw event data
This change adds some of the endpoints for listing raw
event data from the database. It does not yet support
listing events by project id.

It also fixes a problem with the MongoDB driver returning
Mongo's ObjectId instances in the results of the event
query, which makes them impossible to serialize via JSON.

Change-Id: I08d122ecd2f726fb1b2880bc22e28113f6a3aeb1
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>