Use standard CLI options & env vars for creds.
For consistency with the standard openstack CLIs, we use the same
set of config/command line/env var options to provide credentials
for the glance pollster's interaction with the glance-registry.

(The exception here is the ability to pass in a previously acquired
authentication token via --os-auth-token/OS_AUTH_TOKEN, as this
would not make sense for a long-lived service given the limited
lifetime of keystone tokens.)

As well as having the advantage of familiarity, this approach
allows users to avoid encoding sensitive credentials in config
files (instead environment variables or command line options
may be used).

Also, if future non-glance pollsters need authenticated access
to other openstack services, more generic naming of the config
options would be preferable.

Change-Id: I7505c3f668262951c034f36ccd15ce706f06bf0f