Split meter publishing from the global config obj
The meter signing and publishing code was tightly coupled
to the global configuration object, making it more difficult
to reuse in an outside project. This change adds arguments
for the configuration settings actually being used in
those libraries and moves use of the global configuration
object higher up the stack to the manager. It also organizes
the definition of the options related to signing and
publishing metering messages so they can more easily by
registered on other configuration objects.

Change-Id: I8b1df0f228d0c2658841c4b7f7f4527414efc9a6
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>