Fix dependency on anyjson
The main requirements.txt and the one for documentation
had different specification for anyjson.

This change also makes explicit the fact that we
do not require anything other than what is listed
in pip-requires to install. The library function
looks at requirements.txt, too, which we don't have
and should not use.

Change-Id: I8e79ad0803e3a9ebeb2a1dae30b2e5abdbbe8276
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>