Fix notification subscription logic
We need to funnel all messages from all exchanges through
the same function to dispatch them to the handlers. If
we don't do that, we have multiple subscribers,
each configured to be interested in different events
but that do not all receive all events. The result is
events being discarded.

Change-Id: I31daa204c5ee13e8e5f767e0f5e61e2e7eafe624
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>