add max/sum project volume and fix tests
Adds api calls to get the max or sum volume for a specific
project.

Moves the logic for deciding when to skip a test that needs the
Mongo map-reduce feature into a new test db module.

Adds TestConnection class to avoid recreating the underlying
Database object when using mim.  This is necessary because
spidermonkey can only create a limited number of instances per
process.

Change-Id: Idefc01978e41ec823533b7dbb164319a916b155c
Signed-off-by: Steven Berler <steven.berler@dreamhost.com>