Some more doc fixes
There was a spurious = at the end of a line in install.rst that made
the TOC go bad.

Updated the link to the new dev mailing list

areas.rst - testing was mentioning a pre-release, changed to first
version.

Reformated (again) the release notes, hoping it will render better.

Updated the ReleaseNotes file to now point to the rst

Change-Id: Iabf03772498094b0b7e1014abb7fa56547feb636