Add new cpu_util meter recording CPU utilization %.
The existing cpu meter records cumlative time, which is not
the most useful measure in all cases.

We now also estimate the CPU utilization % based on sampling
the time over subsequent polling cycles, and correcting for the
fraction of host cores available to the instance. The usefulness
of the util estimate depends in part on the polling period,
which is assumed will be short (of the order of 1s) when this
measurement is used in anger (for example to drive alarming).

Change-Id: Icd7de1b2cbce12396a82a3f34acb6c39c1120992