Provide a way to disable some plugins
Provide configuration options for the three plugin sets
users may want to manage (compute pollsters, central
pollsters, and notification listeners).

Extend the plugin API so the extension manager can ask
each plugin if it should be enabled. This allows, for
example, the libvirt pollster to be loaded but then
recognize that it should not be used and disable itself.

Addresses bug #1021350

Change-Id: I82da823845ec49c1a93272411c43073bd4954377
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>