Use libvirt num_cpu for CPU utilization calculation.
Previously we used the vcpus count for the corresponding flavor,
but this can be incorrect if the instance type was deleted and
re-created with a different number of cores after the instance
was booted.

Also the total number of CPUs on-host is dropped from the formula
as this is not relevant to the CPU util for the instance.

Change-Id: I23e2ad30696f11f413fd1570fe85c3a12fb0bf1e