Fixes to enable the jenkins doc job to work
Some of our documentation dependencies were not listed in
test-requires, so the venv environment in tox needs to be
updated to include them.

Then the build_sphinx command looks for configuration settings
in setup.cfg, so those need to be added.

When Sphinx runs in this way, it ends up creating
RST files for every module it finds in the source tree
so it can document the internal APIs. Those generated files
can be ignored by git.

Change-Id: I21b5f9c6c4d4201893bb3bf87f0b61a053b968f5
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>