Add sql db option to devstack for ceilometer
Ceilometer uses mongodb as the default storage.
This is one of the patches to support sql db option in devstack for
ceilometer, so users can easily choose and set up sql db environment
for ceilometer with 'stack.sh' by explicitly specifying the following
statement in localrc:

   use_database mysql

where mysql is taken as an example.

If users remove the above statement, mongodb is used and set up
by default.

This patch is for devstack to call dbsync to create ceilometer tables.

Change-Id: Idc9bffd0f389fa9760946b412c39210f60042d90
Signed-off-by: Shane Wang <shane.wang@intel.com>