Make the v2 API date query parameters consistent
Use a DateRange object to specify the old start_timestamp,
end_timestamp, and search_offset parameters in a consistent
way for all of the v2 API calls.

This change addresses bug 1094337 and fixes problems with
running the v2 API tests against the most current version
of WSME, which expects all arguments to be listed on the
controller methods explicitly.

The changes to the test-requires files install the
spidermonkey component, needed for some of the tests
that were being skipped, causing this problem to go
unnoticed for a while.

Change-Id: I4d8677730f228701b0cb85e2ab4294958243e31e
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>