Update nova notifier test after nova change
Nova's code path for cleaning up an instance was modified
to log errors to the database, but our tests don't set
up that part of the schema. Use patch to replace the
function that is trying to log the error.

The instance structure also expects to have a "system_metadata"
key in order to complete the cleanup, so add that to our
test data.

Change-Id: I4bc1d2c8b443a51cef8b13b0627010a6f06c42cf
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>