Update V2 API documentation
This changeset adds a Sphinx extension for auto-generating
much of the documentation for a Pecan/WSME API from the
comments and docstrings in the source code. It also updates
the V2 API to include more documentation for API endpoints
and the data types used in the API, as well as sample data
for generating the JSON and XML examples in the output
documentation.

Change-Id: I1bde7805550aa86e9b64495b5c6034ec328479e5
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>