Decouple the nova notifier from ceilometer code
The move to oslo.config introduced a conflict in the nova notifier
because both nova and ceilometer have copies of the
openstack.common.rpc library and define an option for the AMQP
exchange name for the project. This changeset decouples the notifier
plugin in ceilometer from most of the ceilometer code, to remove that
conflict.

The nova notifier is rewritten to emit a new notification message with
event type `compute.instance.delete.samples` instead of making the old
RPC calls directly to the ceilometer collector, and a notification
listener plugin is provided to convert those messages to sample data.

The notifier implementations are split between the one that worked
in folsom and the one that works in grizzly to maintain backwards
compatibility.

get_hypervisor_inspector() is moved to a location where it can be
imported both by the compute agent manager and the notifier
plugin. The definition of `disabled_compute_pollsters` option is also
moved for the same reason.

The tox configuration is changed to run the nova notifier tests
separately from the other tests, since nose cannot import nova and
ceilometer code in the same process.

bug 1130952

Change-Id: I39ba4564c9c14f09dbdd768d7a83f6940e3942ad
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>