Refactor DB tests
This is a big change but a necessary one. It moves all the database
connection stuff, even for test purpose, in
ceilometer.storage.impl_<backend> rather than redeclaring stuff inside the
tests.db module. That means less code and the possibility to run the API
test against various database backends for free.

Change-Id: I5e2ae2a2628de0d1e105c9d525c510fd356295d4
Signed-off-by: Julien Danjou <julien@danjou.info>