Remove gettext.install from ceilometer/__init__.py
The gettext.install() installs a builtin _() function for translation
purpose. If it is called multiple times, the last call will win and the
translation domain set by the last call will be used.

When the ceilometer.compute.nova_notifier driver used within nova, the
gettext.install() call in ceilometer/__init__.py would change the
previous translation domain set by nova. This is not what we want.

We have to remove this shortcut, and put the gettext.install() call
early on in a top level script.

Instead of directly calling gettext.install(), we leverage the
gettextutils.install() in openstack.common, so we can specifiy the
localedir for ceilometer through the environment variable
CEILOMETER_LOCALEDIR.

This is part of the blueprint gettext-i18n-issue.

Change-Id: Icb2dcfb319778042cf569dcb607f579d1e0fda3a