Ceilometer configuration.rst file not using right param names for logging
configuration.rst file lists all the configuration parameters for ceilometer
configuration. However, for logging, these parameter names should be log_dir
and log_file, but currently it is logdir and logfile.

The table also misses log_config.

There should be a link in configuration.rst file to point a user to the
sample configuration file.

Change-Id: Ic96f303ee7e30faf23c0a6433af5cb5111d35616
fixes: bug #1171905