Added missing source variable in storage drivers
Addresses bug 1163479
*tests/storage/base.py
*ceilometer/storage/models.py
*ceilometer/storage/impl_hbase.py
*ceilometer/storage/impl_mongodb.py
*ceilometer/storage/impl_sqlalchemy.py

Added self.assertIn statements to test_get_resources() and test_get_meters()
functions in /tests/storage/base.py

These assert statements check for the existence of the source variable and
also if the source variable is in the set of sources hard coded by the database
test data.

Added the source variable to models.Resource and models.Meter in the HBase,
MongoDB, and SQLAlchemy drivers.  Added the source variable to the
specifications for the Resource and Meter classes in models.py

Change-Id: I5a407bc8c7ce82a71be7e56705c540ece4573340