Add a direct Ceilometer notifier
This adds a notifier driver for Oslo that will send directly samples to
the collector, avoiding an entire round-trip of the data in their
notification form.

Currently, our only pulisher (ceilometer.publisher.meter) relies on
openstack.common.rpc which doesn't allow cross project RPC calls, so
it's no possible to use this publisher for now. But I'm confident this
limitation will be bypassed soon as we redesign the RPC API.

We will also add other methods to publish soon (HTTP, UDP…) that will be
usable with this notifier driver.

Blueprint: oslo-multi-publisher

Change-Id: Ie639b2fdd3dce24680285ce91f1a024fcfb3b5d0
Signed-off-by: Julien Danjou <julien@danjou.info>