alarm: fix MongoDB alarm id
The current implementation actually fails unit test because it tries to
use alarm id as the internal MongoDB object ID. That doesn't work
because MongoDB object ID must have a certain format and length.

Rather than using a backend specific alarm id, let's just use an uuid
based one that is not tight to the back-end.

Change-Id: I9ea79fc56b8beaf1efceddeaa5e9e43b620e5e98
Signed-off-by: Julien Danjou <julien@danjou.info>