Do the same auth checks in the v2 API as in the v1 API
- Don't limit access to admin.
- Reuse the get_limit_to_project() function from v1/acl.py

This requires some patches to wsme to fix the Response status (so we
can set it to a 401). We don't have this yet, so I am posting an
interim patch. Once we have the next version of wsme I'll fix the
status_code that is returned (see the TODO comments).

Change-Id: Ib39f045b3f9c4ff1f851e13d528fd117aed45c34