Set the default pipline config file for tests
Some of the tests that use a pipeline as a side-effect
(mostly the API tests, now that we create a pipline
automatically in the Pecan hook) only ran if there was
a config file in the production location. This change
resets the default for the pipeline_cfg_file option
to a location that is valid while the tests are running,
even if there is no configuration file in /etc/ceilometer.

Change-Id: I1164df85fd65002660d48d8d76cbfc4a3664f0ea
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>