pipeline: switch publisher loading model to driver
This change modify the loading of the publisher so it's driver based, like
our current storage model. That will allow to use the same publisher
multiple times with different parameters in the future.

This change also adds a new publisher called 'test' that comes from a
factorisation of some of the testing code.

Blueprint: pipeline-publisher-url

Change-Id: Ie26beac213383fff2db759e2df216ba1f454ef9c
Signed-off-by: Julien Danjou <julien@danjou.info>