Capture instance metadata in reserved namespace.
The concept of a reserved namespace for instance metadata
is intended to support the Heat autoscaling use-case.

When spinning up new instances in a autoscaling group, Heat
will set some additional user metadata on the instances to
identify their group membership.

Previously, instance user metadata was not captured in the
metering store and thus could not be queried on.

Now the reserved namespace(s) are used to signal to the
compute agent that some user metadata should not be discarded.
This allows the statistics for a group of associated instances
to be aggregated in a single query.

Change-Id: I838db38bf2e01a0e5e4d76f3f401d7f41e1dff57