Change period type in alarms API to int.
Previously this was float, which looks like an over-sight,
as we will not be supporting sub-second evaluation.

Also float is a mismatch with the v2 statistics API, which
requires that the given period be an int (the alarm period
is passed to the statistics API during threshold eval).

Finally, the sqlalchmey schema defines the period as Integer,
so it makes no sense for the types stored mongo & mysql to
differ.

Submitting this one-liner as a separate patch, as it may be
slightly controversial because it requires that all existing
alarms stored in mongo be deleted and re-created (as we don't
have a schema migration mechnanism for mongo).

Change-Id: Id999005e4de5d77309b0a32ad582804525ec1494