Filter query op:gt does not work as expected
At present, when using timestamp to filter resource or sample data, ceilometer
is using gte for start_timestamp and lt for end_timestamp.
This fix will pass the timestamp operator in the request to the storage level
to support gt operator for the start_timestamp and le for the end_stamp.

The changes will be applied to the get_resources and get_samples methods of
hbase, mongodb and sqlalchemy.

Change-Id: I3fb7651b588400df804ab8bb522084e766744f96
Fixes: Bug 1193926