Transformer to convert between units.
Partially addresses BP transformer-unit

Provide a new transformer to apply scaling conversions, either
replacing or emitted in addition to the original sample.

The target counters are identified either by name or by unit.

The scaling can be expressed as either a straight multiplicative
factor or as a string expression to be eval'd.

Configured as per usual via the pipeline.yaml, for example:

    ...
    transformers:
        - name: "unit_conversion"
          parameters:
              source:
                  unit: "°C"
              target:
                  unit: "°F"
                  scale: "(volume * 1.8) + 32"
              replace: True

Change-Id: I4c3f64376c6f994afd0b1a25dc399daabe2a2ec5