Emit cpu_util from transformer instead of pollster
The emission of the cpu_util counter from the compute CPUUtilPollster
can be superceeded by the RateOfChangeTransformer (the intended
target usercase for the new transformer).

The caching of the CPUStats is now no longer required.

Note that the pipeline.yaml must be updated in deployments, otherwise
the cpu_util meter will no longer be collected.

Change-Id: I79ff047595b960de7a2acb2ec81c357fe2ee21e5