Fixes non-string metadata query issue
Based on current wsme implement, all the metadata query will get
the metadata value as string type. As a result, all the non-string
metadata query will not get correct result. Since the data is store
in database with correct data type. The fix will try to eval the
value before setting so as to get the correct data type.

Fixes bug 1200577

Change-Id: I8114e816da123b9dc08f32f9022db0e1b9fc2e5a