Doc: improve doc about Nova measurements
As explained in bug #1202938, the 'cpu_util' measurement documentation may not
be clear enough. In this patch, we specify that it returns the average
utilisation when there are more than one VCPU (since the measurement is divided
by the number of VCPUS during the 'scale' phase of the transformation).

We also explicitely state that the Nova measurements return information about
the guests, not the hosts.

Change-Id: I69f46b5f82dc0dec2a584ea002dbee19bd55568d