Remove use_greenlets argument to MongoClient
use_greenlets causes gevent to be needed. As per the docs:

"If you need to use standard Python threads in the same process as
Gevent and greenlets, run monkey.patch_socket(), rather than
monkey.patch_all(), and create a MongoClient with use_greenlets=True.
The MongoClient will use a special greenlet-aware connection pool."

We do not use threads, so it seems like we're doing extra complexity
without needing it right now. Also, gevent is not python3 or pypy
compatible, so we'd be bringing in a dependency that would make things
harder moving forward.

Change-Id: I7cd2f5958fdef585cd515c8b1b7ae9fb62c26980