Add a fake UUID to Meter on API level
The database model of Meter does not have a unique key while
Resource has resource_id and Alarm has alarm_id. This key
can be used to identify a meter object by one single value.
The value of meter_id is the combination of resource_id and
couneter_name with base64 encoding. The mainly useage of this
key is pagination query.
It is part of bp/paginate-db-search.

Implements blueprint paginate-db-search

Change-Id: I602d0a370921dcfb02b5bc17c909937c35874e23