Changes FakeMemcache to set token to expire on utcnow + 5 mins
fixes bug 1240994

keystoneclient 0.4 will use utcnow to compare expiration date.
The FakMemcache class sets expiration to datetime.now which sets
date for local timezone. If this timezone happens to be behind
utc the test will fail with expiration error because we are
comparing utc + 5 mins with utc - 1 or more hours.

Change-Id: I551506a5aaf0aef77476b11786399a66225603bf