Replace tests.base part6
It is the sixth step to replace using ceilometer.tests.base with
openstack.common.test.
In this patch is changed for using
openstack.common.test and added needed fixtures in
test classes.
Added in ceilometer.tests.db code from ceilomete.tests.base
because ceilometer.tests.base will not be used in the future.

Change-Id: I3cacca4eff19eae8874decd892ab1225f95cfee9